Generalize fromMetadata
This turned out to be rather involved change. IsMeta become not derivable due to
appearance of `f a' in type with polymorphic f.

Following strategy was used: IsMetaPrim was introduced which is derivable and
elements of Metadata are instances of it. IsMeta instances now could be derived
using anyclass strategy.

This is breaking change. Annoyingly so but perhaps acceptable if generalization
pays off in oka-flow