Big rework of handling extra data
HKD stuff is dropped completely. It turned out to be rather unwieldy and only
`f` we used was completely wrong. In short we added requirements of Applicative
but data type that we have doesn't have sensible Applicative

This means we don't need IsMeta/IsFromMeta split and can use GND/DerivingVia.

We also generalized metadataKeySet in order to give access to IsMetaPrim
dictionaries.