Change semantics of records
Now all data types encoded as null are elided from records and missing fields
are treated as nulls.

This is required in order to work with polymorphic fields. Aeson uses horrible
hacks with IncoherentInstances to make it work. But our instances are intended
to be source of truth so we can be more prescriptive

 1. This allows to add and remove optional field while maintaining validity of
    JSON and keeping hashes stable.

 2. This is another variant of JSON normalization removing encoding ambiguity